backend/requirements.txt
amqp==2.2.2
astroid==1.6.1
autopep8==1.3.4
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
crcmod==1.7
Django==1.11.10
django-cors-middleware==1.3.1
django-environ==0.4.4
django-extensions==1.9.9
django-polymorphic==2.0.2
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
idna==2.6
inflection==0.3.1
isort==4.3.3
jsonfield==2.0.2
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oss2==2.4.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
PyJWT==1.5.3
pylint==1.8.2
python-dateutil==2.7.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
typing==3.6.4
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
wrapt==1.10.11
